Bug 1432679. Do multiplication in nsGIFDecoder2::FinishImageDescriptor as int64_t to avoid overflow. r=aosmond
pixels_remaining is already an int64_t.

UltraBlame original commit: 3509bf5962d123b1ae2064979523f8b567e4847a

+++ image/decoders/nsGIFDecoder2.cpp
@@ -880,7 +880,8 @@ nsGIFDecoder2::FinishImageDescriptor(const char* aData)
   }
 
   
-  mGIFStruct.pixels_remaining = frameRect.Width() * frameRect.Height();
+  mGIFStruct.pixels_remaining =
+    int64_t(frameRect.Width()) * int64_t(frameRect.Height());
 
   if (haveLocalColorTable) {
